Move methods from AxisInlierUtils to InlierUtil
We will need some of those methods later inside another core util class.
